Add handling of $CMUS_HOME variable and clean up code using cmus home
Hardcoding the home directory of cmus to ~/.cmus is fine, but allowing
users to override it using an environment variable - CMUS_HOME - has its
advantages too.  It requires two lines of code to deal with, so it's no
biggie.

This does require that the rest of the code has to be updated to use
cmus_home_dir, which is a good thing, as there wouldn't be much point in
having that variable otherwise.

Signed-off-by: Nikolai Weibull <now@bitwi.se>

[ Timo: Changed main.c to use getenv("CMUS_HOME") instead of linking to
  misc.o and xstrjoin.o ]