seek: Ceil position to duration - 5s
Without this patch if you do

    :player-stop
    :seek 9999m

playing starts but it will not seek at all.  It's slightly more logical
to seek to duration - 5s instead.  The duration - 5s limit is already
used for relative seeking to prevent accidental seeking to end of file
and thus jumping to the next track.